Reduce compile warnings

+++ guava/src/test/scala/magnolify/guava/test/FunnelDerivationSuite.scala
@@ -93,7 +93,7 @@ class BytesSink extends PrimitiveSink {
   }
 
   override def putByte(b: Byte): PrimitiveSink = {
-    oos.writeByte(b)
+    oos.writeByte(b.toInt)
     this
   }
 
@@ -112,7 +112,7 @@ class BytesSink extends PrimitiveSink {
   }
 
   override def putShort(s: Short): PrimitiveSink = {
-    oos.writeShort(s)
+    oos.writeShort(s.toInt)
     this
   }
 
@@ -142,7 +142,7 @@ class BytesSink extends PrimitiveSink {
   }
 
   override def putChar(c: Char): PrimitiveSink = {
-    oos.writeChar(c)
+    oos.writeChar(c.toInt)
     this
   }
